Heading/Paragraph adjustments
This update adjusts the font-size for h4 to scale nicer and look less like
an H3.

It also provides margins to direct `div` siblings of the article container,
to mimic the margins a `<p>` tag would have.

Lastly, the margins of `<p>` and direct `<div>` have been bumped up from
`1em` to `1.5em`.